+++ assets/linux-install.sh
@@ -1,4 +1,5 @@
-curl -L -o mangdl.AppImage https://github.com/MangDL/MangDL/releases/download/3.0.0.4/mangdl-linux-x86_64.AppImage
+#! /bin/sh
+curl -L -o mangdl.AppImage https://github.com/MangDL/MangDL/releases/download/3.0.1.0/mangdl-linux-x86_64.AppImage
 chmod +x ./mangdl.AppImage
 ./mangdl.AppImage --appimage-extract
 rm -rf ./mangdl.AppImage