+++ mangdl/api/providers/mangakomi.py
@@ -1,12 +1,7 @@
-from ..base import soup, urel
-
 scanlator = "Manga Komi"
 base_url = "https://mangakomi.com"
 template = "wordpress"
 
 ch_num_fn = "wmcc"
+rch_fn = "msa"
 src = "data-src"
-
-def rch_fn(url):
-    url = f"{base_url}/manga/{urel(url).parts[2]}/ajax/chapters/"
-    return soup(url, method="post")