+++ mangdl/api/providers/mangaweeaboo.py
@@ -1,0 +1,6 @@
+scanlator = "Weeaboo Scans"
+base_url = "https://mangaweeaboo.com/"
+template = "wordpress"
+
+rch_fn = "msa"
+src = "data-src"