+++ docs/changelog.md
@@ -2,13 +2,32 @@
     CHANGELOG
 </h1>
 
+## **MangDL v.3.0.1.0**
+
+### **ADDED**
+
+- <a target="_blank" href="https://mangaweeaboo.com/">Weeaboo Scans</a> provider
+- Added `msa` function in the `mangdl/api/providers/templates/wordpress.py`'s `rch_fn` templates for the providers to use.
+
+### **CHANGED**
+
+- Changed `mangdl/api/providers/templates/wordpress.py`'s `template.manga` function to check if a provider script declares a `manga_title`, if which it did will pass the `ms` or master soup to it and use the returned value as the title of the manga.
+- In `mangdl/api/providers/firstkiss.py`, changed the `rch_fn`' value to `setsu` and removed the current `rch_fn` function and its imports.
+- In `mangdl/api/providers/mangadex.py` L126, changed the f-string to normal string as it lacks an f-string placeholder.
+- In `mangdl/api/providers/mangakomi.py`, `mangdl/api/providers/mangasushi.py`, `mangdl/api/providers/vinmanga.py`, and `mangdl/api/providers/xunscans.py`, changed the `rch_fn`' value to `msa` and removed the current `rch_fn` function and its imports.
+
+### **FIXED**
+
+- Removed `rch_fn` and its imports in `mangdl/api/providers/paeanscans.py`.
+- Fixed `mangdl/utils/settings.py`'s `readcfg` function to load `yaml` files safely.
+
 ## **MangDL v.3.0.0.4**
 
-**CHANGED**
+### **CHANGED**
 
 - Changed `mangdl/api/providers/templates/wordpress.py`'s `template.manga` function to check if a provider script declares a `manga_title`, if which it did will pass the `ms` or master soup to it and use the returned value as the title of the manga.
 
-**FIXED**
+### **FIXED**
 
 - Removed duplicate `chdls` function in `mangdl/api/providers/mangadex.py`.
 - Fixed Setsu Scans provider by adding a `manga_title` function which returns the appropriate title when given the correct the master soup.
@@ -32,9 +51,14 @@ Attempted to add the `rich` library as a required library and failed.
 
 ## **MangDL v.3.0.0.0**
 
-**CHANGED**
+### **CHANGED**
 
 - Using a modified semantic versioning system, which functions like the old semver but has two major versions for the breaking change visible to the user and the other for developers.
+- Minimized the length of provider scripts further by making templates for `ch_num_fn` and `rch_num_fn` functions, defaulting to `ch_num_fn.breadcrumb` and `rch_num_fn.tdo` functions respectively.
+
+### **FIXED**
+
+- Fixed name inconsistencies for the function `rch_num_fn` which for some instances are named `rch_num_fun`.
 
 ## **MangDL v.2.0.2**
 
@@ -92,7 +116,7 @@ NOW, THIS IS A PROVIDER SPREE!
 ### **CHANGED**
 
 - `cloudflare` must be set to `True` in the provider script if the said provider is protected by cloudflare's UAM. This is in case a method of bypassing it has been found out.
-d
+
 - Use `generic` as a template for providers who does not use a template.
 
 - Removed `Search` dataclass. You can now use `search` function without passing it, instead you pass the arguments directly.